+++ docs/news.rst
@@ -19,6 +19,12 @@ Version updates
 
 .. _releases: https://github.com/4dn-dcic/tibanna/releases
 
+  **Nov 18, 2022** The latest version is now 3.0.0_.
+    - Tibanna now supports AWS Graviton-based instances.
+    - The instance type configuration now allows single instances (e.g., ``t3.micro``) and lists (e.g., ``[t3.micro, t3.small]``). If ``spot_instance`` is enabled, Tibanna will run the workflow on the instance with the highest available capacity. If ``spot_instance`` is disabled, it will run the workflow on the cheapest instance in the list.
+    - The option ``other_instance_types`` for ``behavior_on_capacity_limit`` has been removed. It will fall back to ``wait_and_retry``.
+
+
   **Mar 10, 2022** The latest version is now 2.0.0_.
     - The default Python version for Tibanna is now 3.8 (or 3.7). Python 3.6 is no longer supported.